Cast length of compression buffer to int for 64-bit platforms.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@6718 7a7537e8-13f0-0310-91df-b6672ffda945

+++ filter/rastertolabel.c
@@ -1075,7 +1075,7 @@ PCLCompress(unsigned char *line,	/* I - Line to compress */
   * Set the length of the data and write it...
   */
 
-  printf("\033*b%dW", comp_ptr - CompBuffer);
+  printf("\033*b%dW", (int)(comp_ptr - CompBuffer));
   fwrite(CompBuffer, comp_ptr - CompBuffer, 1, stdout);
 
  /*
